Added new Test cases
