+++ test/test_api.py
@@ -1,17 +1,9 @@
-import pytest
 from flask import request
 
+from test import client
 from app import app
-from app.routes import SERIES_URL, SERIES, SUPPORTED_COLORS, IN_BUILD_IMAGES
+from app.routes import SERIES_URL, SERIES
 from app.routes import get_by_id, finder
-
-
-@pytest.fixture
-def client():
-    """ Initiation of Testing """
-    app.config["TESTING"] = True
-    with app.test_client() as client:
-        yield client
 
 
 def test_home_page(client):
@@ -41,66 +33,3 @@ def test_series_by_id(client):
     for i in SERIES:
         response = client.get(f"/{i}/10")
         assert response.get_json() == get_by_id(i, 10)
-
-
-def test_quote_generator(client):
-    """ Testcase for quote image generator with colored background """
-    response = client.get("/generate/blank?fore=yellow&back=green&size=250")
-    assert request.args["fore"] == "yellow"
-    assert request.args["back"] == "green"
-    assert request.args["size"] == "250"
-    assert response.status_code == 200
-    assert response.headers["Content-Type"] == "image/png"
-
-
-def test_quote_generator_with_in_build_images(client):
-    """ Testcase for quote image generator with image background """
-    for image in IN_BUILD_IMAGES:
-        response = client.get(
-            f"/generate/moneyheist/2/image?src={image}&color=yellow&size=150"
-        )
-        assert request.args["src"] == image
-        assert request.args["color"] == "yellow"
-        assert request.args["size"] == "150"
-        assert response.status_code == 200
-        assert response.headers["Content-Type"] == "image/png"
-
-
-def test_quote_generator_with_user_image(client):
-    """ Testcase for quote image generator with user image background """
-    image = "https://www.gstatic.com/webp/gallery/3.png"
-    response = client.get(
-        f"/generate/moneyheist/5/image?src={image}&color=yellow&size=100"
-    )
-    assert request.args["src"] == image
-    assert request.args["color"] == "yellow"
-    assert request.args["size"] == "100"
-    assert response.status_code == 200
-    assert response.headers["Content-Type"] == "image/png"
-
-
-def test_blank(client):
-    """ Testcase for quote image generator with colored background """
-    response = client.get("/generate/blank")
-    assert response.status_code == 200
-    assert response.headers["Content-Type"] == "image/png"
-
-
-def test_image(client):
-    """ Test case for quote image generator with image background """
-    response = client.get("/generate/image?text=hello+world")
-    assert request.args["text"] == "hello world"
-    assert response.status_code == 200
-    assert response.headers["Content-Type"] == "image/png"
-
-
-def test_color(client):
-    """ Test case for supported colors """
-    response = client.get("/colors")
-    assert SUPPORTED_COLORS == response.get_json()
-
-
-def test_in_buld_images(client):
-    """ Test case for buid-in images """
-    response = client.get("/images")
-    assert IN_BUILD_IMAGES == response.get_json()
